requirements.txt
tensorflow-gpu==2.10.1
keras==2.10.0
numpy==1.23.5
matplotlib==3.5.3
seaborn==0.11.2
pandas==1.4.4
scikit-learn==1.1.3
opencv-python==4.6.0.66
albumentations==1.3.1
pillow==9.2.0
scipy==1.9.3
tqdm==4.64.1
imgaug==0.4.0

